Fix ssl protocol error when sonoff basic tries to connect to the tornado server on debian
This enables TLSv1.1 support for the tornado https server which the sonoff
tries to connect to. TLSv1.1 is insecure and therefore disabled by default
by most modern packages.
But it seems the sonoff firmware can only do TLSv1.1
Without this I got a ssl protocol error and I could not move to FinalStage.